requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pydantic==2.5.2
transformers==4.36.0
torch==2.1.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
langchain==0.1.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
